feat: Merged duplicat kafka code
