+++ golang/cmd/kafka-debug/kafka.go
@@ -2,30 +2,16 @@ package main
 
 import (
 	"github.com/confluentinc/confluent-kafka-go/kafka"
+	"github.com/united-manufacturing-hub/united-manufacturing-hub/internal"
 	"go.uber.org/zap"
 	"time"
 )
 
-func setupKafka(boostrapServer string) (consumer *kafka.Consumer) {
-	configMap := kafka.ConfigMap{
-		"bootstrap.servers": boostrapServer,
-		"security.protocol": "plaintext",
-		"group.id":          "mqtt-debug",
-	}
-	var err error
-	consumer, err = kafka.NewConsumer(&configMap)
-	if err != nil {
-		panic(err)
-	}
-
-	return
-}
-
 func startDebugger() {
-	kafkaConsumerClient.Subscribe("^ia.+", nil)
+	internal.KafkaConsumer.Subscribe("^ia.+", nil)
 	for !ShuttingDown {
 
-		msg, err := kafkaConsumerClient.ReadMessage(5) //No infinitive timeout to be able to cleanly shut down
+		msg, err := internal.KafkaConsumer.ReadMessage(5) //No infinitive timeout to be able to cleanly shut down
 		if err != nil {
 			if err.(kafka.Error).Code() == kafka.ErrTimedOut {
 				continue
