+++ golang/cmd/kafka-init/kafka.go
@@ -1,114 +1,14 @@
 package main
 
 import (
-	"context"
-	"github.com/confluentinc/confluent-kafka-go/kafka"
-	"go.uber.org/zap"
+	"github.com/united-manufacturing-hub/united-manufacturing-hub/internal"
 	"strings"
-	"time"
 )
-
-func setupKafka(boostrapServer string) (adminClient *kafka.AdminClient) {
-	configMap := kafka.ConfigMap{
-		"bootstrap.servers": boostrapServer,
-		"security.protocol": "plaintext",
-		"group.id":          "kafka-init",
-	}
-
-	var err error
-	adminClient, err = kafka.NewAdminClient(&configMap)
-	if err != nil {
-		panic(err)
-	}
-
-	return
-}
 
 func initKafkaTopics(topics string) {
 	topiclist := strings.Split(topics, ";")
 	for _, topic := range topiclist {
-		err := CreateTopicIfNotExists(topic)
+		err := internal.CreateTopicIfNotExists(topic)
 		panic(err)
 	}
-}
-
-func GetMetaData() (metadata *kafka.Metadata, err error) {
-	metadata, err = kafkaAdminClient.GetMetadata(nil, true, 1*1000)
-	return
-}
-
-var lastMetaData *kafka.Metadata
-
-func TopicExists(kafkaTopicName string) (exists bool, err error) {
-	//Check if lastMetaData was initialized
-	if lastMetaData == nil {
-		// Get initial map of metadata
-		lastMetaData, err = GetMetaData()
-		if err != nil {
-			return false, err
-		}
-	}
-
-	//Check if current metadata cache has topic listed
-	if _, ok := lastMetaData.Topics[kafkaTopicName]; ok {
-		zap.S().Debugf("[CACHED] Topic %s exists", kafkaTopicName)
-		return true, nil
-	}
-
-	//Metadata cache did not have topic, try with fresh metadata
-	lastMetaData, err = GetMetaData()
-	if err != nil {
-		return false, err
-	}
-
-	if _, ok := lastMetaData.Topics[kafkaTopicName]; ok {
-		zap.S().Debugf("[CACHED] Topic %s exists", kafkaTopicName)
-		return true, nil
-	}
-
-	return
-}
-
-//goland:noinspection GoVetLostCancel
-func CreateTopicIfNotExists(kafkaTopicName string) (err error) {
-	exists, err := TopicExists(kafkaTopicName)
-	if err != nil {
-		return err
-	}
-	if exists {
-		return
-	}
-
-	var cancel context.CancelFunc
-	defer func() {
-		if cancel != nil {
-			cancel()
-		}
-	}()
-	topicSpecification := kafka.TopicSpecification{
-		Topic:         kafkaTopicName,
-		NumPartitions: 1,
-	}
-	var maxExecutionTime = time.Duration(5) * time.Second
-	d := time.Now().Add(maxExecutionTime)
-	var ctx context.Context
-	ctx, cancel = context.WithDeadline(context.Background(), d)
-	topics, err := kafkaAdminClient.CreateTopics(ctx, []kafka.TopicSpecification{topicSpecification})
-	if err != nil || len(topics) != 1 {
-		zap.S().Errorf("Failed to create Topic %s : %s", kafkaTopicName, err)
-		return
-	}
-
-	select {
-	case <-time.After(maxExecutionTime):
-		zap.S().Errorf("Topic creation deadline reached")
-		return
-	case <-ctx.Done():
-		err = ctx.Err()
-		if err != nil && err != context.DeadlineExceeded {
-			zap.S().Errorf("Failed to await deadline: %s", err)
-			return
-		}
-	}
-	return
 }
