+++ golang/cmd/kafka-to-blob/kafka.go
@@ -6,119 +6,13 @@ import (
 	"encoding/base64"
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 	"github.com/minio/minio-go/v7"
+	"github.com/united-manufacturing-hub/united-manufacturing-hub/internal"
 	"go.uber.org/zap"
 	"time"
 )
 
-func setupKafka(boostrapServer string) (consumer *kafka.Consumer, producer *kafka.Producer, adminClient *kafka.AdminClient) {
-	configMap := kafka.ConfigMap{
-		"bootstrap.servers": boostrapServer,
-		"security.protocol": "plaintext",
-		"group.id":          "kafka-to-blob",
-	}
-
-	var err error
-	consumer, err = kafka.NewConsumer(&configMap)
-	if err != nil {
-		panic(err)
-	}
-
-	producer, err = kafka.NewProducer(&configMap)
-	if err != nil {
-		panic(err)
-	}
-
-	adminClient, err = kafka.NewAdminClient(&configMap)
-	if err != nil {
-		panic(err)
-	}
-
-	return
-}
-
-var lastMetaData *kafka.Metadata
-
-func TopicExists(kafkaTopicName string) (exists bool, err error) {
-	//Check if lastMetaData was initialized
-	if lastMetaData == nil {
-		// Get initial map of metadata
-		lastMetaData, err = GetMetaData()
-		if err != nil {
-			return false, err
-		}
-	}
-
-	//Check if current metadata cache has topic listed
-	if _, ok := lastMetaData.Topics[kafkaTopicName]; ok {
-		zap.S().Debugf("[CACHED] Topic %s exists", kafkaTopicName)
-		return true, nil
-	}
-
-	//Metadata cache did not have topic, try with fresh metadata
-	lastMetaData, err = GetMetaData()
-	if err != nil {
-		return false, err
-	}
-
-	if _, ok := lastMetaData.Topics[kafkaTopicName]; ok {
-		zap.S().Debugf("[CACHED] Topic %s exists", kafkaTopicName)
-		return true, nil
-	}
-
-	return
-}
-
-func GetMetaData() (metadata *kafka.Metadata, err error) {
-	metadata, err = kafkaAdminClient.GetMetadata(nil, true, 1*1000)
-	return
-}
-
-//goland:noinspection GoVetLostCancel
-func CreateTopicIfNotExists(kafkaTopicName string) (err error) {
-	exists, err := TopicExists(kafkaTopicName)
-	if err != nil {
-		return err
-	}
-	if exists {
-		return
-	}
-
-	var cancel context.CancelFunc
-	defer func() {
-		if cancel != nil {
-			cancel()
-		}
-	}()
-	topicSpecification := kafka.TopicSpecification{
-		Topic:         kafkaTopicName,
-		NumPartitions: 1,
-	}
-	var maxExecutionTime = time.Duration(5) * time.Second
-	d := time.Now().Add(maxExecutionTime)
-	var ctx context.Context
-	ctx, cancel = context.WithDeadline(context.Background(), d)
-	topics, err := kafkaAdminClient.CreateTopics(ctx, []kafka.TopicSpecification{topicSpecification})
-	if err != nil || len(topics) != 1 {
-		zap.S().Errorf("Failed to create Topic %s : %s", kafkaTopicName, err)
-		return
-	}
-
-	select {
-	case <-time.After(maxExecutionTime):
-		zap.S().Errorf("Topic creation deadline reached")
-		return
-	case <-ctx.Done():
-		err = ctx.Err()
-		if err != nil && err != context.DeadlineExceeded {
-			zap.S().Errorf("Failed to await deadline: %s", err)
-			return
-		}
-	}
-	return
-}
-
 func processKafkaQueue(topic string, bucketName string) {
-	err := kafkaConsumerClient.Subscribe(topic, nil)
+	err := internal.KafkaConsumer.Subscribe(topic, nil)
 	if err != nil {
 		panic(err)
 	}
@@ -130,7 +24,7 @@ func processKafkaQueue(topic string, bucketName string) {
 			continue
 		}
 		var msg *kafka.Message
-		msg, err = kafkaConsumerClient.ReadMessage(5) //No infinitive timeout to be able to cleanly shut down
+		msg, err = internal.KafkaConsumer.ReadMessage(5) //No infinitive timeout to be able to cleanly shut down
 		if err != nil {
 			if err.(kafka.Error).Code() == kafka.ErrTimedOut {
 				continue
@@ -175,7 +69,7 @@ func pushToMinio(imgBytes []byte, uid string, bucketName string, msg *kafka.Mess
 
 	if err != nil {
 		zap.S().Warnf("Failed to put item into blob-storage: %s", err)
-		kafkaProducerClient.Produce(&kafka.Message{
+		internal.KafkaProducer.Produce(&kafka.Message{
 			TopicPartition: kafka.TopicPartition{
 				Topic:     msg.TopicPartition.Topic,
 				Partition: kafka.PartitionAny,
